refactor: MVC 모델에 따른 클래스 구조 변경
