+++ src/main/java/lotto/model/Calculator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.model;
 
 public class Calculator {
